requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-autoslug==1.9.8
django-filter==2.4.0
django-ipware==3.0.2
django-model-utils==4.1.1
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
urllib3==1.26.5
